Major (breaking) update
In order, Indices and Registers attempt to save in nested path, then to given self.path, then to self.source, and locally as a last resort. This fixes a lot of problems.

Added parent procedure in __init__ and append

Starting to experiment with a base_qoid.py to be inherited by qoid.py classes.

The glossary function is bae